Move name mapping to rest client
