+++ gumbo-rest-client/gumbo_rest_client/name_mapping.json
@@ -253,12 +253,5 @@
 		"sequence_max_corr":"SequenceMaxCorr", 
 		"screen_passes_qc":"ScreenPassesQC", 
 		"screen_id":"ScreenID"
-	}, 
-	"depmap_model_type":{
-		"depmap_code":"DepmapCode", 
-		"oncotree_code":"OncotreeCode", 
-		"lineage":"Lineage",
-		"subtype":"Subtype",
-		"primary_disease":"PrimaryDisease"
 	}
 }
