+++ gumbo-rest-client/tests/utils_test.py
@@ -1,4 +1,4 @@
-from gumbo_dao import utils
+from gumbo_rest_client import utils
 import pandas as pd
